Fix hard fault error on measurement
- The biggest cause of all errors came from ridiculous mistakes. No
validation was done during the invoke of the measurement callback, which
caused a segmentation fault.
- Another error is due to the incomplete implementation of the vsnprintf
function in the nano spec, which causes miscalculated characters to
contaminate the stack, making it difficult to track.
- Luckily for now, I've solved the urgent error to get the measurement
data, but currently the uEmbedded library's fslist is not implemented
correctly, so I'm using STL's list for timer implementation. You should
fix the malfunction of the fslist and optimize the behavior of the
SetTimerFromIRQ function.